mod: CacheProvider를 일반 객체로 변경
contexts/
- 불필요해진 Context 삭제
