+++ client/src/App.tsx
@@ -1,4 +1,3 @@
-import { CacheProvider } from 'contexts';
 import { RouterProvider } from 'react-router-dom';
 
 import router from 'router';
@@ -6,11 +5,7 @@ import router from 'router';
 import './css/index.css';
 
 function App() {
-  return (
-    <CacheProvider>
-      <RouterProvider router={router} />
-    </CacheProvider>
-  );
+  return <RouterProvider router={router} />;
 }
 
 export default App;
