--- client/src/contexts/CacheProvider/hooks/index.ts
@@ -1,1 +1,0 @@
-export { default as useCacheActions } from './useCacheActions';
