--- client/src/contexts/index.ts
@@ -1,1 +1,0 @@
-export { default as CacheProvider } from './CacheProvider';
